# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Nutrition tracking API.\n\nError code catalog:\n- See docs/openapi-error-codes.md in repository root.",
        "title": "Nutrition API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/progress/daily": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-goals"
                ],
                "summary": "Get daily progress against user goals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DailyProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/recipes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "List recipes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by recipe name (case-insensitive, partial match)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.RecipeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Create recipe",
                "parameters": [
                    {
                        "description": "Recipe payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRecipeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.RecipeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/recipes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Get recipe by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RecipeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Update recipe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recipe update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateRecipeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RecipeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/user-goals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-goals"
                ],
                "summary": "Get current user goals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserGoalResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-goals"
                ],
                "summary": "Create or update user goals",
                "parameters": [
                    {
                        "description": "User goals payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpsertUserGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserGoalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateRecipeRequest": {
            "type": "object",
            "properties": {
                "ingredients": {
                    "description": "Ingredient list used for nutrition calculation.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RecipeIngredientRequest"
                    }
                },
                "name": {
                    "description": "Human-readable recipe name.",
                    "type": "string",
                    "example": "Rice Bowl"
                },
                "yield_weight_g": {
                    "description": "Final cooked yield weight in grams.",
                    "type": "number",
                    "example": 200
                }
            }
        },
...
        "dto.UpdateRecipeRequest": {
            "type": "object",
            "properties": {
                "ingredients": {
                    "description": "Optional full replacement of ingredient list.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RecipeIngredientRequest"
                    }
                },
                "name": {
                    "description": "Optional recipe name.",
                    "type": "string",
                    "example": "Updated Rice Bowl"
                },
                "yield_weight_g": {
                    "description": "Optional final cooked yield weight in grams.",
                    "type": "number",
                    "example": 210
                }
            }
        },
        "dto.UpsertUserGoalRequest": {
            "type": "object",
            "properties": {
                "activity_level": {
                    "description": "Optional activity level: sedentary|light|moderate|active|very_active.",
                    "type": "string",
                    "example": "moderate"
                },
                "target_carbs_g": {
                    "description": "Daily carbs target in grams.",
                    "type": "number",
                    "example": 220
                },
                "target_fat_g": {
                    "description": "Daily fat target in grams.",
                    "type": "number",
                    "example": 70
                },
                "target_kcal": {
                    "description": "Daily kcal target.",
                    "type": "number",
                    "example": 2200
                },
                "target_protein_g": {
                    "description": "Daily protein target in grams.",
                    "type": "number",
                    "example": 150
                },
                "weight_goal_kg": {
                    "description": "Optional body weight goal in kilograms.",
                    "type": "number",
                    "example": 80
                }
            }
        },
        "handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Stable machine-readable error code.",
                    "type": "string",
                    "example": "invalid_food_payload"
                },
                "message": {
                    "description": "Human-readable error message.",
                    "type": "string",
                    "example": "invalid food payload"
                },
                "request_id": {
                    "description": "Request ID for log correlation.",
                    "type": "string",
                    "example": "f328f2f67798/SrQv8xTxwA-000001"
                }
            }
        },
...
        "handlers.DailyProgressResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Target date in YYYY-MM-DD.",
                    "type": "string",
                    "example": "2026-02-17"
                },
                "remaining_carbs_g": {
                    "description": "Remaining carbs grams (target - total).",
                    "type": "number",
                    "example": 20
                },
                "remaining_fat_g": {
                    "description": "Remaining fat grams (target - total).",
                    "type": "number",
                    "example": 10
                },
                "remaining_kcal": {
                    "description": "Remaining kcal (target - total).",
                    "type": "number",
                    "example": 350
                },
                "remaining_protein_g": {
                    "description": "Remaining protein grams (target - total).",
                    "type": "number",
                    "example": 30
                },
                "target_carbs_g": {
                    "description": "Target carbohydrate grams for the day.",
                    "type": "number",
                    "example": 220
                },
                "target_fat_g": {
                    "description": "Target fat grams for the day.",
                    "type": "number",
                    "example": 70
                },
                "target_kcal": {
                    "description": "Target kcal for the day.",
                    "type": "number",
                    "example": 2200
                },
                "target_protein_g": {
                    "description": "Target protein grams for the day.",
                    "type": "number",
                    "example": 150
                },
                "total_carbs_g": {
                    "description": "Aggregated carbohydrate grams for the day.",
                    "type": "number",
                    "example": 200
                },
                "total_fat_g": {
                    "description": "Aggregated fat grams for the day.",
                    "type": "number",
                    "example": 60
                },
                "total_kcal": {
                    "description": "Aggregated kcal for the day.",
                    "type": "number",
                    "example": 1850
                },
                "total_protein_g": {
                    "description": "Aggregated protein grams for the day.",
                    "type": "number",
                    "example": 120
                }
            }
        },
...
        "handlers.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.APIError"
                }
            }
        },
...
        "handlers.RecipeIngredientResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Creation timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "food_id": {
                    "description": "Referenced food ID.",
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "description": "Ingredient row ID.",
                    "type": "integer",
                    "example": 1
                },
                "position": {
                    "description": "Optional ordering position.",
                    "type": "integer",
                    "example": 1
                },
                "raw_weight_g": {
                    "description": "Raw ingredient weight in grams.",
                    "type": "number",
                    "example": 200
                },
                "recipe_id": {
                    "description": "Parent recipe ID.",
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "description": "Last update timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                }
            }
        },
        "handlers.RecipeResponse": {
            "type": "object",
            "properties": {
                "carbs_per_100g": {
                    "description": "Carbohydrate grams per 100g.",
                    "type": "number",
                    "example": 28
                },
                "created_at": {
                    "description": "Creation timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "fat_per_100g": {
                    "description": "Fat grams per 100g.",
                    "type": "number",
                    "example": 0.3
                },
                "id": {
                    "description": "Recipe ID.",
                    "type": "integer",
                    "example": 1
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.RecipeIngredientResponse"
                    }
                },
                "kcal_per_100g": {
                    "description": "Energy in kcal per 100g.",
                    "type": "number",
                    "example": 130
                },
                "name": {
                    "description": "Recipe name.",
                    "type": "string",
                    "example": "Rice Bowl"
                },
                "protein_per_100g": {
                    "description": "Protein grams per 100g.",
                    "type": "number",
                    "example": 2.7
                },
                "updated_at": {
                    "description": "Last update timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "yield_weight_g": {
                    "description": "Final cooked yield weight in grams.",
                    "type": "number",
                    "example": 200
                }
            }
        },
        "handlers.UserGoalResponse": {
            "type": "object",
            "properties": {
                "activity_level": {
                    "description": "Optional activity level.",
                    "type": "string",
                    "example": "moderate"
                },
                "created_at": {
                    "description": "Creation timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T08:00:00Z"
                },
                "id": {
                    "description": "User goal ID.",
                    "type": "integer",
                    "example": 1
                },
                "target_carbs_g": {
                    "description": "Target carbs grams per day.",
                    "type": "number",
                    "example": 220
                },
                "target_fat_g": {
                    "description": "Target fat grams per day.",
                    "type": "number",
                    "example": 70
                },
                "target_kcal": {
                    "description": "Target kcal per day.",
                    "type": "number",
                    "example": 2200
                },
                "target_protein_g": {
                    "description": "Target protein grams per day.",
                    "type": "number",
                    "example": 150
                },
                "updated_at": {
                    "description": "Last update timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T08:00:00Z"
                },
                "user_id": {
                    "description": "Owner user ID.",
                    "type": "integer",
                    "example": 1
                },
                "weight_goal_kg": {
                    "description": "Optional weight goal in kilograms.",
                    "type": "number",
                    "example": 80
                }
            }
        },
...
    }
}